Move FreeListT out to separate file.
