+++ vespamalloc/src/vespamalloc/malloc/CMakeLists.txt
@@ -4,6 +4,7 @@ vespa_add_library(vespamalloc_malloc OBJECT
     malloc.cpp
     allocchunk.cpp
     common.cpp
+    freelist.cpp
     mmappool.cpp
     threadproxy.cpp
     memblock.cpp
@@ -18,6 +19,7 @@ vespa_add_library(vespamalloc_mallocd OBJECT
     mallocd.cpp
     allocchunk.cpp
     common.cpp
+    freelist.cpp
     mmappool.cpp
     threadproxy.cpp
     memblockboundscheck.cpp
@@ -33,6 +35,7 @@ vespa_add_library(vespamalloc_mallocdst16 OBJECT
     mallocdst16.cpp
     allocchunk.cpp
     common.cpp
+    freelist.cpp
     mmappool.cpp
     threadproxy.cpp
     memblockboundscheck.cpp
@@ -49,6 +52,7 @@ vespa_add_library(vespamalloc_mallocdst16_nl OBJECT
     mallocdst16_nl.cpp
     allocchunk.cpp
     common.cpp
+    freelist.cpp
     mmappool.cpp
     threadproxy.cpp
     memblockboundscheck.cpp
