+++ vespamalloc/src/vespamalloc/malloc/datasegmentdst.cpp
@@ -1,8 +1,8 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-#include <vespamalloc/malloc/datasegment.hpp>
-#include <vespamalloc/malloc/memblockboundscheck_dst.h>
+#include "datasegment.hpp"
+#include "memblockboundscheck_dst.h"
 
-namespace vespamalloc {
+namespace vespamalloc::segment {
 
 template class DataSegment<MemBlockBoundsCheck>;
