+++ vespamalloc/src/vespamalloc/malloc/freelist.cpp
@@ -1,9 +1,8 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-#include "datasegment.hpp"
-#include "memblockboundscheck_d.h"
+#include "freelist.hpp"
 
 namespace vespamalloc::segment {
 
-template class DataSegment<MemBlockBoundsCheck>;
+template class FreeListT<0x40000>;
 
 }
